Server/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
sqlparse==0.5.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
whitenoise==6.4.0
